move graph and other file movement
