+++ docs/bokeh/source/_images/gallery/node_and_edge_attributes.png
