+++ docs/bokeh/source/docs/reference/plotting.rst
@@ -7,9 +7,7 @@ The ``bokeh.plotting`` API is Bokeh's **primary** interface, and lets you focus
 on relating glyphs to data. It automatically assembles plots with default
 elements such as axes, grids, and tools for you.
 
-Below is a basic example:
-
-.. bokeh-plot:: docs/user_guide/examples/concepts_plotting.py
+.. bokeh-plot:: __REPO__/examples/bars/basic.py
     :source-position: above
 
 This reference guide chapter has a few sections split out for the different
