+++ examples/geo/choropleth.py
