+++ examples/geo/eclipse.py
