+++ examples/geo/eclipse_data/upath17.shp
