+++ examples/graph/customize.py
@@ -1,9 +1,8 @@
 import math
 
-from bokeh.io import output_file, show
 from bokeh.models import Ellipse, GraphRenderer, StaticLayoutProvider
 from bokeh.palettes import Spectral8
-from bokeh.plotting import figure
+from bokeh.plotting import figure, show
 
 N = 8
 node_indices = list(range(N))
@@ -31,5 +30,4 @@ graph.layout_provider = StaticLayoutProvider(graph_layout=graph_layout)
 
 plot.renderers.append(graph)
 
-output_file('graph.html')
 show(plot)
