+++ examples/graph/interaction.py
@@ -1,12 +1,11 @@
 import networkx as nx
 
-from bokeh.io import output_file, show
-from bokeh.models import (BoxSelectTool, Circle, HoverTool, MultiLine,
-                          NodesAndAdjacentNodes, Plot, Range1d, TapTool)
+from bokeh.models import (BoxSelectTool, Circle, EdgesAndLinkedNodes, HoverTool,
+                          MultiLine, NodesAndLinkedEdges, Plot, Range1d, TapTool)
 from bokeh.palettes import Spectral4
-from bokeh.plotting import from_networkx
+from bokeh.plotting import from_networkx, show
 
-G=nx.karate_club_graph()
+G = nx.karate_club_graph()
 
 plot = Plot(width=400, height=400,
             x_range=Range1d(-1.1,1.1), y_range=Range1d(-1.1,1.1))
@@ -24,10 +23,9 @@ graph_renderer.edge_renderer.glyph = MultiLine(line_color="#CCCCCC", line_alpha=
 graph_renderer.edge_renderer.selection_glyph = MultiLine(line_color=Spectral4[2], line_width=5)
 graph_renderer.edge_renderer.hover_glyph = MultiLine(line_color=Spectral4[1], line_width=5)
 
-graph_renderer.selection_policy = NodesAndAdjacentNodes()
-graph_renderer.inspection_policy = NodesAndAdjacentNodes()
+graph_renderer.selection_policy = NodesAndLinkedEdges()
+graph_renderer.inspection_policy = EdgesAndLinkedNodes()
 
 plot.renderers.append(graph_renderer)
 
-output_file("interactive_graphs.html")
 show(plot)
