+++ examples/graph/interaction_edgeslinkednodes.py
@@ -1,12 +1,11 @@
 import networkx as nx
 
-from bokeh.io import output_file, show
 from bokeh.models import (BoxSelectTool, Circle, EdgesAndLinkedNodes,
                           HoverTool, MultiLine, Plot, Range1d, TapTool)
 from bokeh.palettes import Spectral4
-from bokeh.plotting import from_networkx
+from bokeh.plotting import from_networkx, show
 
-G=nx.karate_club_graph()
+G = nx.karate_club_graph()
 
 plot = Plot(width=400, height=400,
             x_range=Range1d(-1.1,1.1), y_range=Range1d(-1.1,1.1))
@@ -29,5 +28,4 @@ graph_renderer.inspection_policy = EdgesAndLinkedNodes()
 
 plot.renderers.append(graph_renderer)
 
-output_file("interactive_graphs.html")
 show(plot)
