+++ examples/graph/node_and_edge_attributes.py
